requirements.txt
africastalking==1.2.7
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contextlib2==21.6.0
convex==0.6.0
fastapi==0.110.0
frozenlist==1.4.1
h11==0.14.0
idna==3.6
multidict==6.0.5
pydantic==2.6.3
pydantic_core==2.16.3
PyJWT==1.7.1
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
schema==0.7.5
signalwire==2.1.0
six==1.16.0
sniffio==1.3.1
starlette==0.36.3
tenacity==8.2.3
twilio==6.54.0
typing_extensions==4.10.0
urllib3==2.2.1
uvicorn==0.28.0
yarl==1.9.4
